requirements.txt
Django>=3.2,<4.0
beautifulsoup4>=4.9.3
requests>=2.26.0
pytz>=2021.1

# Testing dependencies
pytest>=7.0.0
pytest-django>=4.5.0
pytest-cov>=4.0.0
pytest-benchmark>=4.0.0
pytest-mock>=3.10.0
factory-boy>=3.2.0
coverage>=7.0.0
locust>=2.15.0